refactor(reconcile): split planner.ts into focused modules
Codacy's analyzer keeps stitching consecutive function declarations
together and attributing the metrics of the run to whichever
method-name-shaped token it grabs (e.g. labeling 23 params / cc 16 as
"Method Set" because `new Set()` is the closest thing to a method call
on that line). Splitting the 520-line planner.ts into per-concern
modules gives each function a clear file boundary.

- internal/envs.ts: EnvKey/EnvSet types, ENVLESS sentinel, envSorter,
  envSetsEqual.
- internal/instance.ts: InstanceState, buildInstances, the desired-
  bindings collection, attachMovedFrom, instanceKey/stableStringify.
- internal/rename.ts: resolveRenames + the two-pass match logic
  (movedFrom, then envCoverage shape match) and rename-commit helpers.
- planner.ts: trimmed to the public planReconciliation entry plus the
  per-instance diff/leaf-action emission.

Also swap a couple of `Set<string>` book-keeping structures to
`Map<string, true>` to remove the bare `Set` token Codacy was
mis-parsing.

Behavior unchanged; all 158 cli tests pass.
